Configure CORS, create env var util, revrieve hello world w/axios
This way of doing CORS is less hacky

This env var file will be good for prod and allows me to reusably
set the api paths regardless of the environment. This is akind to
the structure of the services used in my CSCi 3601 projects, or
my URS internship project.

Hits backend with axios to get a hello world with a red border.

Addresses #4 (GET a 'hello world' from Express)
Project: 'Establish reachable Express backend'