Render a view with a GET from Mongoose, from a POSTed req
Rather large commit, but all of the pieces need to be here to setup
the infrastructure of creating and pulling in an item.

I connect to a mongo database on MongoDB Atlas and get its
collection via Mongoose. I have a `~/config` folder that points to
the right env var depending on the environment.

I made a simple 'Ratings' model that just takes a user and a rating.
It can only be POSTed to via something like Insomnia, but it can be
GET from via a hook in the frontend.

Addresses #5 (GET from Mongoose)
Project: 'Establish reachable Express backend