Convert a rating's number into a star rating system
This was a fun commit, because it caused my browser to lock up from
dependency recurrsion. It took about two minutes to realize that
my `Rating FC` and the MUI `Rating` component were competing for
that namespace.

Speaking of the MUI `Rating`, it is from MUI's Lab category of
components. As per the MUI docs on the lab: https://material-ui.com/components/about-the-lab/
'This package hosts the incubator components that are not yet ready
to move to the core'. They are subject to breaking changes and take
less priority than core components. However, they rely on usage
to get an idea about demand and problems, so this helps to use it.
So far it seems like a nice component that does its job.

Issue #28 (Convert Repos and Ratings into MUI card grid)
Project: 'Breakdown `Cockpit` into subcomponents'