Add google login (TERRIBLE COMMIT!)
This commit is terrible because it is so large and I don't have
time to explain.

Basically, google oauth with passport.js took me an embrassasingly
long time to configure to get perfectly. I have worked with google
oauth in an internship and agile project based class before, and
even wrote docs on it. BUT I find that the neccisary ways to configure
it in the full stack are extremely brittle. I like to push working
commits, so I am not breaking this up into smaller files.

Because of factors like this being vanilla JS, where I learned how
to do google oauth in an express+react course and how dated it is,
and contrasting online examples that are outdated, this is a patchwork
of a setup. I don't have the time to sufficiently give it a rundown
or comment....

BUT basically,
1) `Login FC` invokes a `user` context that calls `/auth/google` and
I can get profile info there. Once the user is directed to the login page
and logs in, the client is kicked back to a callback and that redirects back
home.
2) If a user is new, their info is grabbed and stored as a mongoose
`user`. If not new, it just returns their info.
3) Client now has user info.

There is a mess of proxy stuff I need to explain later, that's its whole
thing.

This is rather last second on my deadline for this code, so I have
to move on.

Addresses #35 (Initialize Passport.js)
Addresses #36 (Implement Google OAuth strategy for Passport)
Project: 'Implement Google OAuth w/Passport.js'