Add `requireLogin` middleware, implement in ratings
The middleware is passed to each backend route's.

Frontend requests that require login pass in `{withCredentials: true}`
to each request to send the user cookie.

As long as the user is logged in, they pass the middleware and the
request works.

Addresses #51 (Require login for POSTs/DELETEs)
Project: 'Implement Google Oauth w/Passport.js'